Add game events

+++ include/Connect4/ui/elements/buttonGroup.h
@@ -22,6 +22,6 @@ void C4_UI_ButtonGroup_FreeResources(C4_UI_ButtonGroup* stack);
 void C4_UI_ButtonGroup_Destroy(C4_UI_ButtonGroup* stack);
 void C4_UI_ButtonGroup_SetButtonIndex(C4_UI_ButtonGroup* stack, size_t buttonIndex, SDL_Renderer* renderer, const char* str, C4_FontType font, float ptSize);
 void C4_UI_ButtonGroup_Draw(C4_UI_ButtonGroup* stack, SDL_Renderer* renderer);
-long long C4_UI_ButtonGroup_HandleMouseEvents(C4_UI_ButtonGroup* stack, SDL_Event* event, SDL_Renderer* renderer);
+int C4_UI_ButtonGroup_HandleMouseEvents(C4_UI_ButtonGroup* stack, SDL_Event* event, SDL_Renderer* renderer);
 void C4_UI_ButtonGroup_TransformResize(C4_UI_ButtonGroup* stack, const SDL_FRect rect);
 void C4_UI_ButtonGroup_CenterInWindow(C4_UI_ButtonGroup* stack, C4_Axis axis);
